extract: collapse download helper into utils/texts (no separate file)
Reviewer feedback — the helper exists to map an extension to a label
that already lives in utils/texts, so putting it in its own
components/other/downloadLabel.ts was an extra hop. Lift
downloadLabelByExt + getDownloadLabel into utils/texts next to
buttonsTitles, derive buttonsTitles.downloadPdf/Gdb/GeoJson from the
map so they stay exposed for the existing referrers, and delete the
intermediate file. Both downloader components import from one place.
